backend/pyproject.toml
[tool.poetry]
name = "blogsite"
version = "0.1.0"
description = ""
authors = ["PavelLexmaul <pavel.lexmaul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg = "^3.1.10"
psycopg2 = "^2.9.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
django-seed = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
